Move changes into original Browser component
